Initial code

+++ src/Boxfusion.Health.HealthCommon.Core/Domain/BackBoneElements/Enum/RefListAllergyIntoleranceSeverities.cs
@@ -6,13 +6,25 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.BackBoneElements.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "AllergyIntoleranceSeverities")]
-	public enum RefListAllergyIntoleranceSeverities: int
+	public enum RefListAllergyIntoleranceSeverities: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Mild")]
 		mild = 1,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Moderate")]
 		moderate = 2,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Severe")]
 		severe = 3
 	}
