+++ src/Boxfusion.Health.HealthCommon.Core/Domain/BackBoneElements/Enum/RefListSNOMEDCTRouteCodes.cs
@@ -6,7 +6,7 @@ using System.Text;
 namespace Boxfusion.Health.HealthCommon.Core.Domain.BackBoneElements.Enum
 {
 	[ReferenceList("Fhir", "SNOMEDCTRouteCodes")]
-	public enum RefListSNOMEDCTRouteCodes: int
+	public enum RefListSNOMEDCTRouteCodes: long
 	{
 	}
 }
