+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Cdm/Enum/RefListAreaTypeCategory.cs
@@ -6,11 +6,23 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Cdm.Enum
 {
+    /// <summary>
+    /// 
+    /// </summary>
 	[ReferenceList("Cdm", "AreaTypeCategory")]
-	public enum RefListAreaTypeCategory : int
+	public enum RefListAreaTypeCategory : long
 	{
+        /// <summary>
+        /// 
+        /// </summary>
         Municipality = 1,
+        /// <summary>
+        /// 
+        /// </summary>
         Region = 2,
+        /// <summary>
+        /// 
+        /// </summary>
         Province = 3
     }
 }
