+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Cdm/Enum/RefListProcedureCodingSystem.cs
@@ -6,11 +6,21 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Cdm", "ProcedureCodingSystem")]
-	public enum RefListProcedureCodingSystem: int
+	public enum RefListProcedureCodingSystem: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Free text")]
 		freeText = 1,
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("SNOMED")]
 		snomed = 2
 	}
