+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListCompositionStatuses.cs
@@ -6,15 +6,33 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "CompositionStatuses")]
-	public enum RefListCompositionStatuses: int
+	public enum RefListCompositionStatuses: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Preliminary")]
 		premilinary = 1,
+		
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Final")]
 		final = 2,
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Amended")]
 		amended = 3,
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Entered in Error")]
 		enteredInError = 4
 	}
