+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListDocumentRelationshipTypes.cs
@@ -6,15 +6,33 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "DocumentRelationshipTypes")]
-	public enum RefListDocumentRelationshipTypes: int
+	public enum RefListDocumentRelationshipTypes: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Replaces")]
 		replaces = 1,
+		
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Transforms")]
 		transforms = 2,
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Signs")]
 		signs = 3,
+
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Appends")]
 		appends = 4
 	}
