+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListImmunisationFundingSources.cs
@@ -6,11 +6,20 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "ImmunisationFundingSources")]
-	public enum RefListImmunisationFundingSources: int
+	public enum RefListImmunisationFundingSources: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Private Funding")]
 		privateFunding = 1,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Public Funding")]
 		publicFunding = 2
 	}
