+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListImmunisationOriginCodes.cs
@@ -6,15 +6,30 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "ImmunisationOriginCodes")]
-	public enum RefListImmunisationOriginCodes: int
+	public enum RefListImmunisationOriginCodes: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Other Provider")]
 		otherProvider = 1,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Written Record")]
 		writtenRecord = 2,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Parent/ Guardian/ Patient Recall")]
 		parentGuardianPateintRecall = 3,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("School Record")]
 		schoolRecord = 4
 	}
