+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListImmunisationProgramEligibilities.cs
@@ -6,11 +6,20 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "ImmunisationProgramEligibilities")]
-	public enum RefListImmunisationProgramEligibilities: int
+	public enum RefListImmunisationProgramEligibilities: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Not Eligible")]
 		notEligible = 1,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Uninsured")]
 		unInsured = 2
 	}
