+++ src/Boxfusion.Health.HealthCommon.Core/Domain/Fhir/Enum/RefListImmunisationReasonCodes.cs
@@ -6,11 +6,20 @@ using System.Text;
 
 namespace Boxfusion.Health.HealthCommon.Core.Domain.Fhir.Enum
 {
+	/// <summary>
+	/// 
+	/// </summary>
 	[ReferenceList("Fhir", "ImmunisationReasonCodes")]
-	public enum RefListImmunisationReasonCodes: int
+	public enum RefListImmunisationReasonCodes: long
 	{
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Procedure to Meet Occupational Requirement")]
 		procedureToMeetOccupationalRequirement = 1,
+		/// <summary>
+		/// 
+		/// </summary>
 		[Description("Travel Vaccinations")]
 		travelVaccinations = 2
 	}
